feat: add landing page
